Move project hooks to folder
